requirements.txt
openai
python-dotenv
Pillow
PyMuPDF
pandas
mammoth
openpyxl
msgraph-sdk